Added copyright headers
According to the commit log, work on the checking functionality started
in 2002 (commit 820dca1912dd3100233342d916c2136334a20180), so I used
that as a start for Axel's line. I couldn't find any information on the
internet about copyright when refactoring, but I figured that adding my
name as well would be appropriate.